# libfetchers: avoid re-copying substituted inputs