@getchoo 2025-09-24
Took a different approach by just copying the fix from https://github.com/NixOS/nix/pull/12911 directly here instead of trying to re-use the whole code path. Seems to have the same effect!