@getchoo 2025-09-25
Added a release note as well, since it seems this bug has a bit of a following 😄