@getchoo 2025-09-25
While #14050 is an obvious improvement to `fetchToStore()` caching generally, AFAICT it doesn't fix this specific issue (tested by merging it into 0175f7e8365074356497d132451870c9eae9240f and running `nix eval` on this repo again, same as above)

This sorta makes sense? As it only touches the unsubstituted code path there, plus the explicit caching originally introduced in #12911 [remains](https://github.com/NixOS/nix/pull/14050/files#diff-05e2127dd673970e2253ad7844aac6fc33a3cdbe5057dd2e0d9714b2fc20c8ecR144-R154)

Now I don't really know much about libfetchers yet (or how Nix works generally...), so take this with a grain of salt: I'm pretty sure this patch is still necessary since the accessor we're constructing here hasn't undergone any previous fingerprinting or received an entry to the cache. This ends up similar to what the regular `getAccessor` methods for each `InputScheme` does, which is `ensure the content exists in the store -> fingerprint it -> create an entry in the cache for that fingerprint`. The unsubstituted code path handles this by simply calling to those `getAccessor` methods which handle all the caching for us, but since here we're grabbing an accessor from the store directly, we need to implement that same caching logic (similar to how we already implement the second step of fingerprinting)